LayoutTests/fast/js/kde/cast-expected.txt

PASS Boolean(1) === true is true
PASS var s = String; s(1) === '1' is true
PASS n = Number; n(true) === 1 is true
PASS String(Array('a', 'b'	)) is 'a,b'
PASS successfullyParsed is true
